requirements.txt
psutil>=5.9.0
py-cpuinfo>=8.0.0
GPUtil>=1.4.0
PySide6>=6.4.0
pywin32>=305; platform_system == "Windows"
pyinstaller>=5.7.0
